==> courses - workshops/ALERTAS EMPRESA/0_CRAWLER_.ipynb <==
#!/usr/bin/env python
# coding: utf-8

import os
import re
import string
import time
from datetime import datetime, timedelta
import shutil
import glob
import IPython

import pandas as pd
import unicodedata
from pyspark.sql.types import *
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.support.select import Select
from sqlalchemy import create_engine
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By

!pip install pytesseract

from PIL import Image

import pytesseract

## Obtener Imagen Captcha
Este método se encarga guardar la screenshot y recortarla para que solo quede el captcha.

def get_captcha(driver, element, path):
    location = element.location
    size = element.size
    driver.save_screenshot(path)

    image = Image.open(path) #abre el screenshot
    #define vertices para recortar la imagen con las letras del captcha
    left = location['x'] +5
    top = location['y'] +5
    right = location['x'] + size['width']+5
    bottom = location['y'] + size['height'] + 5

    image = image.crop((left, top, right, bottom))  # recorta la screen en la parte donde se encuentra la img del captcha
    image.save(path, 'png')  # 

## Procesar Imagen
este método consiste en procesar la imagen del captcha para transformarla a blanco y negro y que sea más facil el reconocimiento de letras

def procesar_img():
    img = Image.open('captcha2.png')
    img = img.convert("RGBA")
    pixdata = img.load()
    
    img2 = img
    size2 = 100, 100
    img2.thumbnail(size2, Image.ANTIALIAS)
    img2.save("input-con-fe.png")
    pixdata2 = img2.load()
    #cambiar letras de captcha a negro para facilitar su transformación a string
    for y in range(img2.size[1]):
        for x in range(img2.size[0]):
            if not (pixdata2[x,y][0] >230 and pixdata2[x,y][1] >230 and pixdata2[x,y][2] >230): #fondo de captcha blanco
                pixdata2[x,y] = (0,0,0,255)
            else:
                pixdata2[x,y] = (255,255,255,255)
    img2.save("input-con-fe-black.png")

def captcha_to_string_4():
    
    pytesseract.pytesseract.tesseract_cmd = r"C:\Program Files (x86)\Tesseract-OCR\tesseract.exe"
    image = Image.open("input-con-fe-black.png")
    texto_captcha = pytesseract.image_to_string(image, config="-c tessedit_char_whitelist=ABCDEFGHIJKLMNOPQRSTUVWXYZ -psm 6") #solo considere letras mayus
    texto_captcha = texto_captcha.replace(' ', '') #Elimina espacios entre las letras
    return texto_captcha

def pasar_captcha(driver, ruc):
    #pasar captcha sunat
    
    driver.get("http://e-consultaruc.sunat.gob.pe/cl-ti-itmrconsruc/frameCriterioBusqueda.jsp")
    img = driver.find_element_by_xpath('/html/body/form/table/tbody/tr/td/table[2]/tbody/tr[1]/td[5]/img')
    get_captcha(driver, img, "captcha2.png")
    procesar_img()
    texto1 = captcha_to_string_4().upper()
    #print(texto1)
    
    element = driver.find_element_by_xpath('//*[@id="s1"]/input')
    element.send_keys(ruc) #ingresa ruc a consultar
    
    element = driver.find_element_by_xpath('/html/body/form/table/tbody/tr/td/table[2]/tbody/tr[1]/td[6]/input')
    element.send_keys(texto1) #ingresa código captcha
    
    element = driver.find_element_by_xpath('/html/body/form/table/tbody/tr/td/table[2]/tbody/tr[1]/td[7]/input')
    element.click() #presiona buscar
    
    
    for handle in driver.window_handles:
        driver.switch_to_window(handle)
        

    #print("veamos")
    boton_reactiva = driver.find_element_by_xpath('//*[@id="div_estrep"]/table/tbody/tr[2]/td[3]/form/input[1]')
    boton_reactiva.click()
    #print("ahora")
    
    check = driver.find_element_by_xpath('//*[@id="tieneCoactiva2"]')
    #print(dir(check))
    #driver.close()
    return check.get_property('checked')
    #attrs = driver.execute_script('var items = {}; for (index = 0; index < arguments[0].attributes.length; ++index) { items[arguments[0].attributes[index].name] = arguments[0].attributes[index].value }; return items;', check)
    #print(attrs)
    #try:    
    #    alert = driver.switch_to.alert
    #    alert.accept() #acepta los popup de confirmacion en caso que aparezcan
    #except:
    #    return
    


class SunatCrawler(object):
    def __init__(self, proxy=False):
        #self.hoy = datetime(2020, 9, 2, 13, 0, 0)
        self.hoy = datetime.today()
        self.hoy_format = self.hoy.strftime('%d/%m/%Y')
        self.ayer = self.hoy - timedelta(days=3)
        self.ayer_format = self.ayer.strftime('%d/%m/%Y')
        self.default_path = "C:\\Users\\B38381\\Desktop\\COFIDE\\"
        
        #self.engine = create_engine("mysql://jquiza:7841wQnGTO1CovvE@108.59.81.200:3306/ibk_business_account")
        #self.connection = self.engine.connect()
        self.current_path = os.path.join(os.getcwd(), 'selenium')
        print("PATH A BUSCAR DRIVER: ", self.current_path)
        
        options = Options()
        #options.add_argument("--incognito")
        options.add_argument('--no-sandbox')
        options.add_argument("--allow-running-insecure-content")
        
        if proxy:
            PROXY = "34.216.89.174:53281"
            options.add_argument('--proxy-server=%s' % PROXY)
            options.add_argument('--disable-software-rasterizer')
            #options.add_argument('--headless')
            options.add_argument("--disable-web-security")
            print(PROXY)
        else:
            options.add_argument('--no-proxy-server')  #  agregado 01/09/2020 VMVP (evita interrupciones)
        
        #options.add_argument("--start-fullscreen")
        user_agent = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.132 Safari/537.36'
        options.add_argument('user-agent={}'.format(user_agent))
        options.add_experimental_option("prefs", {
            "download.default_directory": r"{0}".format(self.current_path),
            "download.prompt_for_download": False,
            "download.directory_upgrade": True,
            "safebrowsing.enabled": True
        })
        

        self.driver_path = os.path.join(self.current_path, 'chromedriver.exe') #os.path.join(self.current_path, 'utils', 'chromedriver.exe')
        #self.url = 'http://s340vp22:9080/web/guest/home?p_p_id=58&p_p_lifecycle=0&p_p_state=maximized&p_p_mode=view&saveLastPath=false'
        self.driver = webdriver.Chrome(self.driver_path, options=options)


#def get_clear_browsing_button(driver):
#    return driver.find_element_by_css_selector('* /deep/ #clearBrowsingDataConfirm')
#
#
#def clear_cache(driver, timeout=10):
#    driver.get('chrome://settings/clearBrowserData')
#
#    wait = WebDriverWait(driver, timeout)
#    wait.until(get_clear_browsing_button)
#
#    get_clear_browsing_button(driver).click()
#
#    wait.until_not(get_clear_browsing_button)

#sunat = SunatCrawler()
#pasar_captcha(sunat.driver, '10451391459')

## ACUMULACION DE CASOS FALTANTES

import pandas as pd
from market_conecction import connection_market

from querys_especificos import query_deuda
from querys_cofide import query_validaciones

# CSV 
path_0 = 'scraper_0.csv'

scraper = pd.read_csv(
    path_0, dtype={'RUC': str}, parse_dates=['Fecha']
).sort_values(
    by='Fecha', ascending=False
).drop_duplicates(
    subset=['RUC'], keep='first'
).reset_index(
    drop=True
)


scraper.to_csv(path_0, index=False)
scraper = scraper[
    scraper['RUC'].notnull()
]
print(scraper.shape)
scraper.head()

scraper.tail()

# --> AQUI CAMBIAR

acum = True
rucs = set(
  ['20529527609','20554505954', '20601966027', '20455162531']
)


if acum:
    # QUERY A BUSCAR RUCS
    query_asignado = """
        select DISTINCT(NUM_RUC) 
        from ODS.HD_SEGUIMIENTO_SUBASTA_ETAPAS a    
        where 1 = 1
        --and FECHA_CORTE = (SELECT MAX(FECHA_CORTE) FROM ODS.HD_SEGUIMIENTO_SUBASTA_ETAPAS)
        and TASA > 0
        and NUM_RUC is not null
    """
    data_asignado = pd.read_sql(
       query_asignado , connection_market
    )
    data_asignado['NUM_RUC'] = data_asignado['NUM_RUC'].astype(str)

    rucs = set(data_asignado['NUM_RUC']) 
    
    print(len(rucs), type(rucs))
    del data_asignado

### >> LOS QUE FALTAN

rucs = rucs - set(scraper['RUC'])
len(rucs), type(rucs)

import time

if acum:
    connection_market.close()

from datetime import datetime
lista_add_df = []

sunat = SunatCrawler(proxy=False)
for orden, ruc in enumerate(list(rucs), start=1):

        intentos = 0
        while True:
            try:
                intentos += 1
                response = int(not(pasar_captcha(sunat.driver, ruc)))
                sunat.driver.delete_all_cookies()    
                sunat.driver.refresh()
                break
            except Exception as e:
                print("", end='.')

                if intentos < 5:
                    sunat.driver.delete_all_cookies()    
                    sunat.driver.refresh()
                elif intentos == 5:
                    sunat.driver.close()
                    sunat.driver.quit()
                    sunat = SunatCrawler(proxy=True)
                    time.sleep(1)
                else:
                    time.sleep(10)
                    sunat.driver.close()
                    sunat.driver.quit()
                    sunat = SunatCrawler(proxy=False)
                    sunat.driver.stop_client()
                    sunat.driver.start_client()
                    intentos = 0
        
        if response in [0, 1] and ruc and str(ruc).lower() != 'none':
            lista_add_df.append(
                ['Victor', str(ruc), response, None, datetime.now()]
            )
            print(str(ruc) + '-' + str(response), end=' ')
        
        if orden % 50 == 0:
            df = pd.DataFrame(lista_add_df, columns=list(scraper.columns))
            df.to_csv(path_0, index=False, header=False, mode='a')
            lista_add_df = []
            
            print("\n GUARDADO \n")
            
        #sunat.driver.close()
        #sunat.driver.quit()


if lista_add_df:
    df = pd.DataFrame(lista_add_df, columns=list(scraper.columns))
    df.to_csv(path_0, index=False, header=False, mode='a')

print(dir(sunat.driver))

print(help(sunat.driver.set_network_conditions))
