==> courses - workshops/ALERTAS EMPRESA/0.1 WORKERS RESCUE.ipynb <==
import gc
import time
import pymssql

import pandas as pd
import numpy as np

def clean_str_to_float(val):
    if pd.isnull(val):
        return None
    
    _str = str(val).replace(" ", "").replace(",", "").replace(".", "")
    
    if not _str.isnumeric():
        print(_str, end=' ')
    
    return int(float(_str)) if  _str.isnumeric() else None

try:
    connection_market.close()
except:
    pass


#conneciom MARKET
connection_market = pymssql.connect(host='10.11.12.90\BDT',
                             user='ibetlmarket',
                             port=1433,
                             password='m@rk3t2o15',
                             database='MARKET'
                            )
time.sleep(1)
gc.collect()

sql_rucs_scraping = """
    SELECT * FROM ODS.MM_SUNAT_CANT_TRABAJADORES 
"""
sunat_crawler = pd.read_sql(
   sql_rucs_scraping, connection_market
).sort_values(by=['FECHA_CARGA'], ascending=False)
print(sunat_crawler.shape)

connection_market.close()
sunat_crawler.head()

sunat_crawler = sunat_crawler.groupby(['PERIODO', 'RUC']).agg(
    {'NRO_TRABAJADORES': 'max', 'NRO_PENSIONISTAS': 'max', 'NRO_PRESTADORES_SERVICIO': 'max'}
).reset_index()

sunat_crawler.shape

sunat_crawler.head()

sunat_crawler.dtypes

### VALIDACION DE DATOS NULOS

rucs_scraping = list(sunat_crawler['RUC'].unique())
print(len(rucs_scraping))

print(sunat_crawler.shape)
sunat_crawler = sunat_crawler[sunat_crawler['NRO_TRABAJADORES'].notnull()]
print(sunat_crawler.shape)

sunat_crawler['ANIO'] = sunat_crawler['PERIODO'].apply(lambda _: _[:4]).astype(int)
sunat_crawler['MES'] = sunat_crawler['PERIODO'].apply(lambda _: _[4:]).astype(int)
sunat_crawler = sunat_crawler.rename(columns={
    'NRO_TRABAJADORES': 'TOTAL_TRABAJADORES',
    'NRO_PENSIONISTAS': 'PENSIONISTA',
    'NRO_PRESTADORES_SERVICIO': 'PRESTADOR_SERVICIO'
})[['PERIODO', 'ANIO', 'MES', 'TOTAL_TRABAJADORES', 'PENSIONISTA', 'PRESTADOR_SERVICIO', 'RUC']]

sep = '__'
sunat_crawler['COMBINACION'] = sunat_crawler['PERIODO'] + sep +  sunat_crawler['RUC']
sunat_crawler.head()

sunat_crawler = sunat_crawler[sunat_crawler['ANIO'] == 2019]
sunat_crawler.head()

csv_name = "cantidad_trabajadores.csv"
cant_trab = pd.read_csv(csv_name, dtype={'RUC': str, 'PERIODO': str})

rucs_registrados = list(cant_trab['RUC'].unique())
combinaciones_registradas = list(cant_trab['COMBINACION'].unique())

print(len(rucs_registrados))
print(len(combinaciones_registradas))
print(cant_trab.shape)

cant_trab.head()

sunat_crawler = sunat_crawler[
    (sunat_crawler['RUC'].isin(rucs_registrados)) & (~sunat_crawler['COMBINACION'].isin(combinaciones_registradas))
]
print(sunat_crawler.shape)

sunat_crawler['PERIODO'].value_counts().sort_index()

pivoteo = pd.pivot_table(
    sunat_crawler, 
    values='TOTAL_TRABAJADORES', index=['RUC'], columns=['PERIODO'], aggfunc=np.max, fill_value=-1
).reset_index()

pivoteo.head()

tiene_menos_1 = pivoteo[
    (pivoteo['201910'] != -1) & (pivoteo['201911'] == -1) &  (pivoteo['201912'] == -1)
]
tiene_menos_1

len(list(tiene_menos_1['RUC'].unique()))

sunat_crawler = sunat_crawler[~sunat_crawler['RUC'].isin(list(tiene_menos_1['RUC'].unique()))]
sunat_crawler['PERIODO'].value_counts()

cant_trab.shape

cant_trab = pd.concat(
    [cant_trab, sunat_crawler], axis=0, ignore_index=True
)
cant_trab['PERIODO'].value_counts().sort_index()

cant_trab = cant_trab[cant_trab['PERIODO'] != '2020']
cant_trab.shape

cant_trab.shape

cant_trab.to_csv(csv_name, index=False)

cant_trab.dtypes

cant_trab['MES'] = cant_trab['MES'].astype(int)

cant_trab.head()

cant_trab['PERIODO'].value_counts().sort_index()

cant_trab['ANIO'].value_counts().sort_index()

cant_trab['MES'].value_counts()
