==> courses - workshops/ALERTAS EMPRESA/0.2 NRO TRABAJADORES.ipynb <==
## VICTOR DICE:
import os
import gc
import time
import requests
import subprocess
import random as rn
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta

import socks
import socket
import pymssql
import numpy as np
import pandas as pd

SEED = int(time.time() / 1000)
os.environ['PYTHONHASHSEED'] = str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

### SOLO ES NECESARIO CORRER UNA VEZ

#!pip install socks

#!pip install pymssql

### OPCIONES

alternativo = 0
priorizado = 1

print(dir(socks))

print(socks.getdefaultproxy())

now = datetime.now()

desfase_2_meses =  datetime.strptime((now - relativedelta(months=2)).strftime('%Y-%m'), '%Y-%m').strftime('%Y%m')
desfase_2_meses

letras = "abcdefghijklmnopqrstuvwyz"
user_aletorio = ''.join([rn.choice(letras) for _ in range(8)])
pass_aleatorio = ''.join([rn.choice(letras) for _ in range(8)])

user_aletorio, pass_aleatorio

users_agents = [
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.182 Safari/537.36",
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.105 Safari/537.36"
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:85.0) Gecko/20100101 Firefox/85.0",
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.182 Safari/537.36 Edg/88.0.705.74",
    "Mozilla/5.0 CK={} (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko",
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36",
    "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/72.0.3626.121 Safari/537.36",
    "Mozilla/5.0 (Windows NT 10.0; WOW64; Trident/7.0; rv:11.0) like Gecko"
]

url = "http://api.sunat.binvoice.net/consulta.php"

headers = {
     'content-type': "application/x-www-form-urlencoded",
     'cache-control': "no-cache",
     'User-Agent': rn.choice(users_agents),
     'Referer': "http://api.sunat.binvoice.net/",
     'Origin': "api.sunat.binvoice.net",
     'Upgrade-Insecure-Requests': '0'
}

#import asyncdns, asyncio

try:
    connection_market.close()
except:
    pass


#conneciom MARKET
connection_market = pymssql.connect(host='10.11.12.90\BDT',
                             user='ibetlmarket',
                             port=1433,
                             password='m@rk3t2o15',
                             database='MARKET'
                            )
time.sleep(1)
gc.collect()

### QUERY COMPLETO > 16 000 CLIENTES

sql_all_rucs_clientes = """
    SELECT DISTINCT(NUM_RUC) FROM ODS.HM_VPC_CLIENTE
    WHERE 1 = 1
    and PERIODO >= '202007'
    and NUM_RUC LIKE '20%'
    and (
        ZONAL in ('BEL ZONAL 1','BEL ZONAL 2','BEL ZONAL 3','BEP ZONAL 1',
                  'BEP ZONAL 2','BEP ZONAL 3', 'BC GRUPO 2', 'BC GRUPO 1')
        OR
        BANCA in ('BE','BEL','BEP', 'BC')
    )
"""

### QUERY PRIORIZADO ALERTAS CARTERA  < 3000

sql_priorizado = """
    SELECT DISTINCT(NUM_RUC)FROM ODS.HD_VMVP_BASE_ALERTA_CARTERA_BE
    WHERE 1 = 1
    and FUNNEL_ALERTA = 1
    and ZONAL IS NOT NULL
    and ZONAL != ''
    and PERIODO >= '202007'
"""

rucs = pd.read_sql(
   sql_priorizado if priorizado else sql_all_rucs_clientes, connection_market
)
print(rucs.shape)

connection_market.close()
rucs.head()

rucs = list(rucs['NUM_RUC'].unique())
len(rucs)

### IDENTIFICACION quienes no registran trabajadores

csv_not = 'not_workers.csv'
not_workers = pd.read_csv(csv_not, dtype={'RUC': str}, parse_dates=['FECHA']).sort_values(
    by=['FECHA'], ascending=False).drop_duplicates(subset=['RUC'], keep='first')

### Vuelve a revisar los detectados hace 2 meses
not_workers = not_workers[
    not_workers['FECHA'] > now - relativedelta(months=2)
]

not_workers.to_csv(csv_not, index=False)
not_workers = pd.read_csv(csv_not, dtype={'RUC': str}, parse_dates=['FECHA'])
not_workers.head()

not_workers.dtypes

list_not = list(not_workers['RUC'].unique())
len(list_not)

### AUTO limpiado de duplicados

csv_name = "cantidad_trabajadores.csv"
cant_trab = pd.read_csv(csv_name, dtype={'RUC': str, 'PERIODO': str})
print(cant_trab.shape)

cant_trab = cant_trab.drop_duplicates(subset=['PERIODO', 'RUC'], keep='last')
cant_trab.to_csv(csv_name, index=False)
print(cant_trab.shape)

### DETETCCIÓN de clientes a buscar actualización

cant_trab = pd.read_csv(csv_name, dtype={'RUC': str, 'PERIODO': str, 'ESTADO': str})
print(cant_trab.shape)
cant_trab.head()

cant_trab.dtypes

cant_trab['PERIODO'].value_counts(dropna=False)

clientes_nuevos = set(rucs) - set(cant_trab['RUC'])
len(clientes_nuevos)

clientes_a_actualizar = set(cant_trab['RUC']) - set(cant_trab[cant_trab['PERIODO'] == desfase_2_meses]['RUC'])
len(clientes_a_actualizar)

sep = '__'
clientes_update = list(
    (clientes_nuevos | clientes_a_actualizar) - set(list_not)
)
del clientes_nuevos
del clientes_a_actualizar

len(clientes_update)

rucs_periodo_existentes = list(cant_trab['COMBINACION'].unique())
del cant_trab

#len(rucs_periodo_existentes)

gc.collect()

rn.shuffle(clientes_update)

if alternativo:
    print("OPEN")
    sproc = subprocess.Popen("C:\\Users\\Villacorta\\Desktop\\Tor Browser\\Browser\\firefox.exe")

agents_tor = [
    'Mozilla/5.0 (Windows NT 10.0; rv:78.0) Gecko/20100101 Firefox/78.0'
]

headers_tor = {
    'content-type': "application/x-www-form-urlencoded",
    'User-Agent': rn.choice(agents_tor),
}

def get_response(ruc):
    user_aletorio = ''.join([rn.choice(letras) for _ in range(8)])
    pass_aleatorio = ''.join([rn.choice(letras) for _ in range(8)])
    
    socks.set_default_proxy(socks.SOCKS5, "127.0.0.1", 9150, True, user_aletorio, pass_aleatorio)
    socket.socket = socks.socksocket
    
    headers_tor.update(
        {
            'User-Agent': rn.choice(agents_tor),
            'X-Forwarded-For': requests.get('http://ifconfig.me/ip').text
        }
    )
    #print()
    
    print(headers_tor)
    response = requests.request("POST", url, data="nruc={}".format(ruc), headers=headers_tor)
    #print(response)
    return response

def get_response_simple(ruc):
    headers.update(
        {'User-Agent': rn.choice(users_agents)}
    )
    response = requests.request("POST", url, data="nruc={}".format(ruc), headers=headers)
    #print(response)
    return response

for ruc in clientes_update:
    response = get_response(ruc) if alternativo else  get_response_simple(ruc)
    try:
        data = response.json()
    except:
        print(" {} ".format(response.status_code))
        time.sleep(60)
        gc.collect()
        continue
    
    if not data.get("success"):
        print(" er ", data.get("message", ""), end=' ')
        time.sleep(60)
        gc.collect()
        continue
    
    estado = data.get('result',{}).get('estado', '')
    print(estado)
    cantidad_trabajadores = data.get('result',{}).get('cantidad_trabajadores',{})
    cantidad_trabajadores = pd.DataFrame(cantidad_trabajadores) if len(cantidad_trabajadores) > 0 else None
    
    ### SI NO TIENE TRABAJADORES REGISTRAODS, lo guardamos aparte para no buscarlo en otras corridas
    if not isinstance(cantidad_trabajadores, pd.core.frame.DataFrame):
        print(ruc, 
              " nt ", end='')
        
        pd.DataFrame(
            [[ruc, datetime.now()]]
        ).to_csv(
            csv_not, 
            mode='a',
            header=False, 
            index=False
        )
        
        gc.collect()
        continue
    
    ### FORMATO
    cantidad_trabajadores.columns = [col.upper() for col in cantidad_trabajadores.columns]
    cantidad_trabajadores['PERIODO'] = cantidad_trabajadores['PERIODO'].apply(lambda _: _.replace('-', '').replace('_', ''))
    cantidad_trabajadores['RUC'] = ruc
    cantidad_trabajadores['COMBINACION'] = cantidad_trabajadores['PERIODO'] + sep +  cantidad_trabajadores['RUC']
    cantidad_trabajadores['ESTADO'] = estado
    ### GUARDADO, y acumulación de los satisfactorios
    cantidad_trabajadores[
        ~cantidad_trabajadores['COMBINACION'].isin(rucs_periodo_existentes)
    ].to_csv(
        csv_name, 
        mode='a',
        header=False, 
        index=False
    )
    print(ruc, end=' ')


if alternativo:
    sproc.kill()
