==> courses - workshops/ALERTAS EMPRESA/0.3 up SUNAT TO SQL.ipynb <==
import pandas as pd

csv_name = "cantidad_trabajadores.csv"
cant_trab = pd.read_csv(csv_name, dtype={'RUC': str, 'PERIODO': str})
print(cant_trab.shape)

cant_trab = cant_trab.drop_duplicates(subset=['PERIODO', 'RUC'], keep='last')
cant_trab.to_csv(csv_name, index=False)
print(cant_trab.shape)

cant_trab.head()

cant_trab.columns = [col.upper().strip() for col in cant_trab.columns]

cant_trab.dtypes

cant_trab['PERIODO'].value_counts(dropna=False)

def get_nro_trab_num(val):
    pre = val.replace(' ', '').split('.')[0]
    
    if pre.isnumeric():
        return int(pre)
    
    try:
        return int(pre)
    except:
        print(pre, end=' ')
        return None
    
    return None


cant_trab['TOTAL_TRABAJADORES'] = cant_trab['TOTAL_TRABAJADORES'].apply(lambda _: get_nro_trab_num(_))
cant_trab['TOTAL_TRABAJADORES'].value_counts(dropna=False).sort_index()

cant_trab['PENSIONISTA'] = cant_trab['PENSIONISTA'].apply(lambda _: get_nro_trab_num(_))

cant_trab['PRESTADOR_SERVICIO'] = cant_trab['PRESTADOR_SERVICIO'].apply(lambda _: get_nro_trab_num(_))

#cant_trab[cant_trab['TOTAL_TRABAJADORES'].notnull()]['TOTAL_TRABAJADORES'].value_counts(dropna=False).sort_index()

from datetime import datetime

cant_trab['FECHA_CARGA'] = datetime.now()
cant_trab = cant_trab.rename(columns={
    'TOTAL_TRABAJADORES': 'NRO_TRABAJADORES',
    'PENSIONISTA': 'NRO_PENSIONISTAS',
    'PRESTADOR_SERVICIO': 'NRO_PRESTADORES_SERVICIO'    
})

cant_trab.head()

cols_up = ['RUC', 'PERIODO', 'NRO_TRABAJADORES', 'NRO_PENSIONISTAS', 'NRO_PRESTADORES_SERVICIO', 'FECHA_CARGA']
cant_trab[cols_up].dtypes


cant_trab[cant_trab['RUC'] == '20100340438'].sort_values(by=['PERIODO'])

cant_trab['PERIODO'].value_counts()

%%time

from sqlalchemy import create_engine
engine = create_engine(r"mssql+pymssql://ibetlmarket:m@rk3t2o15@10.11.12.90\BDT/MARKET")
connection_x = engine.connect()

try:
    cant_trab[cols_up].to_sql(
        name='MM_SUNAT_CANT_TRABAJADORES',
        schema='ODS',
        con=connection_x, 
        index=False, 
        if_exists='replace'
    )
    connection_x.close()
    correct = 1

except Exception as e:
    print("ERROR: ", str(e))
    correct = 0

print(correct)

import pymssql
print("Inciando actualizacion de tblas ODS")
connection = pymssql.connect(
    host='10.11.12.90\BDT',
    user='ibetlmarket',
    port=1433,
    password='m@rk3t2o15',
    database='MARKET'
)

with connection.cursor() as cursor:
    try:
        print("/"*50)
        alter_text_files = """
        ALTER TABLE ODS.MM_SUNAT_CANT_TRABAJADORES ALTER COLUMN RUC VARCHAR(50);
        ALTER TABLE ODS.MM_SUNAT_CANT_TRABAJADORES ALTER COLUMN PERIODO VARCHAR(50) ;
        """
        cursor.execute(alter_text_files)
    except Exception as e:
        print("ERROR: ", str(e))
        
try:
    connection.commit()
    connection.close()
    print("Conexion database cerrada")
except Exception as e:
    print("Error:  ", str(e))

print(cant_trab.shape)
cant_trab = cant_trab[cant_trab['TOTAL_TRABAJADORES'].notnull()]
print(cant_trab.shape)

cant_trab['PERIODO'].value_counts(dropna=False).sort_index()

pivoteo = pd.pivot_table(
    cant_trab, index=['RUC'], columns=['PERIODO'], values='TOTAL_TRABAJADORES', fill_value=-1
)
del pivoteo['202101']
del pivoteo['201910']
del pivoteo['201911']
del pivoteo['201912']

pivoteo['MESES_VALIDOS'] = pivoteo.apply(lambda _: len([val for val in _ if val >= 0]), axis=1)
pivoteo.head()

pivoteo['MESES_VALIDOS'].value_counts().sort_index()

pivoteo[pivoteo['MESES_VALIDOS'] == 11]
