==> courses - workshops/ALERTAS EMPRESA/0.9 FEVE AJUSTADO.ipynb <==
import glob
import pandas as pd

relacion = []

for name in glob.iglob('FEVE_REAJUSTADO_BANCA_EMPRESA_*.xlsx'):
    relacion.append(
        (name, name.split('.')[0].split('_')[-1] )
    )
    print(relacion[-1])

todo = []

for _excel, periodo in relacion:
    print("/"*100 , periodo)
    col_seguir_covid = 'SEGUIR COVID'
    
    # lectura de cada excel
    data = pd.read_excel(_excel, dtype={
        'CU': str, 'CLIENTE': str,  col_seguir_covid: str
    }, engine='openpyxl')
    
    # formato a las columnas y selccino las de interes
    data.columns = [_.upper() for _ in data.columns]
    data = data[['CU', 'CLIENTE', col_seguir_covid]]
    
    
    # formato homogeneo al conntenido, filtrado y seteado decolumnas de apoyo
    for col in data.columns:
        data[col] = data[col].fillna('').apply(lambda _: _.upper())
        
    data = data[data[col_seguir_covid] == 'SI']
    data['PERIODO'] = str(periodo)
    data['SEGUIR_COVID'] = 1
    
    # Acumulado de info con los demás periodos
    print(data.shape)
    todo.append(data)
    display(data.head(3))
    
    del data

data = pd.concat(
    todo, axis=0
)
data['PERIODO'].value_counts()

data.head()

data.shape

data[data['CU'] == '0011872710']

data.to_excel("FEVE_SEGUIR_COVID.xlsx", index=False)

%%time

from sqlalchemy import create_engine
engine = create_engine(r"mssql+pymssql://ibetlmarket:m@rk3t2o15@10.11.12.90\BDT/MARKET")
connection_x = engine.connect()

try:
    data.to_sql(
        name='HM_FEVE_REAJUSTADO_SEGUIR_COVID',
        schema='ODS',
        con=connection_x, 
        index=False, 
        if_exists='replace'
    )
    connection_x.close()
    correct = 1
except Exception as e:
    print("ERROR: ", str(e))
    correct = 0

print(correct)

connection_x.close()
