==> courses - workshops/ALERTAS EMPRESA/2.1_MEDIANA_BIVARIADO.ipynb <==
# Nativos
import random as rn
import os
import sys
import gc

#calculo
import numpy as np
import pandas as pd
import scipy

#grafico
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
%matplotlib inline
sns.set(style="whitegrid")

#warning ignore future
import warnings
# warnings.simplefilter(action='ignore', category=FutureWarning)
warnings.filterwarnings("ignore")
gc.collect()

SEED = 29082013
os.environ['PYTHONHASHSEED'] = str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

from utils_alertas import *

filtro_banca_mediana

#!pip install pandas==0.25

#!pip install git+git://github.com/shichenxie/scorecardpy.git

#nativos
import os
import unicodedata

#terceros
import pandas as pd
#import pymssql
#import missingno as msno
from collections import Counter
import scorecardpy as sc

#grafico
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
%matplotlib inline
sns.set(style="whitegrid")

data = pd.read_csv('AVANCE_5.5.csv', dtype={
    'COD_UNICO': str,
    'PERIODO': str
})
data.head()

print(len(data.columns))

[col for col in data.columns if 'FINAL' in col]

col_target = 'target_FINAL_fusion_atraso_feve_2Mbueno'

data[data['COD_UNICO'] == str(8443816).zfill(10)].sort_values(by=['PERIODO'])[
    ['target_SF_8_dias_atraso_o_mas_1Mbueno']
]

## DATA A ANALIZAR

col_no_review = list(set(['COD_UNICO', 'PERIODO', 'DES_GRUPO_ECONOMICO', 'ENCARGADO', 
                 'NOMBRE_COMPLETO', 'NOMBRE_JEFE', 'RUC', 'GRUPO_ECON_FEVE_SEGUIR', 'MTO_CASH_IN', 'RFMClass_quartil', 'DESC_ATRASO_SF',
                 'GRUPO_ECON_CLASIF_NO_NORMAL', 'GRUPO_ECON_FEVE_DURO', 'RFMClass_septil', 'MTO_CASH_IN', 'DESC_ATRASO_SF',
                 'SUBSECTOR'] + [col for col in data.columns if 'DESCR' in col.upper()]))
col_no_review = list(set(col_no_review))
len(col_no_review)

all_data = data.copy()

print(data.shape)
data = filtro_banca_mediana(data)
print(data.shape)

data['PERIODO'].value_counts(dropna=False)

data['ZONAL'].value_counts(dropna=False)

#### REVISION DE NO NULOS

data[data[col_target].notnull()].shape, data.shape

nn = data[data[col_target].notnull()].isnull().sum()
no_nulos = list(nn[nn == 0].index)
len(no_nulos)

no_nulos.remove('PERIODO')
no_nulos.remove('COD_UNICO')

## Peso de la evidencia  / informacion de su valor

pd.set_option("display.max_rows", 1000)
pd.set_option("display.max_colwidth", 1000)
pd.set_option("display.max_columns", 1000)

#iv_yoy = sc.iv(
#    data[data[col_target].notnull()].drop(
#        [col for col in data.columns if 'target' in col and col != col_target] + ['COD_UNICO', 'PERIODO'], axis=1), y=col_target
#)

#select = iv_yoy[
#    (iv_yoy['info_value'] > 0.1) #&  (iv_yoy['info_value'] <= 1)
#]
#select.shape

#dict_iv = select.set_index('variable').to_dict().get('info_value', {})

#select[:55]

#select[55:110]

#select[110:165]

#select[165:220]

sc.__version__

import conda
conda.__version__

pd.__version__   ### 1.1.3 ---> versión en la que funciona scorerapy (isuue reportado para 1.2.0 en adelante)

if 'FEC_INGRESO' in data.columns:
    del data['FEC_INGRESO']

data['target_FINAL_fusion_atraso_feve_2Mbueno']

## >> BINS POR WOE/ IV

%%time

if creation_mode:
    cortes = sc.woebin(
        data[
            (data[col_target].notnull())
        ].drop(
            [col for col in data.columns if 'target' in col and col != col_target] + col_no_review, 
            axis=1
        ),
        y=col_target,
        method='tree',
        no_cores=4,
        count_distr_limit=0.05  # el 5% por bin es lo mínimo recomendable para arboles
    )

#help(sc.woebin)

#sc.woebin_plot(bins)

if creation_mode:
    dic_maxprob_corte = {}
    for k, v in cortes.items():
        #print(k, type(v))
        ##print(">"*20, ' ', k.upper())
        #display(v)
        #dic_maxprob_corte[k] = v['badprob'].max()
        max_probabilidad_malo = v[v['bin'] != 'missing']['badprob'].max()
        informacion_valor = v['total_iv'].max()
        dic_maxprob_corte[k] = [
            max_probabilidad_malo, informacion_valor, (max_probabilidad_malo + informacion_valor) / 2
        ]

if creation_mode:
    listado_mayor_prob_corte = sorted([(k, v) for k, v in dic_maxprob_corte.items()], key=lambda _: _[1][1], reverse=True)
    listado_mayor_prob_corte[:15]

if creation_mode:
    vars_top = [_[0] for _ in listado_mayor_prob_corte[:500]]
    len(vars_top)

if 'FEC_INGRESO' in no_nulos:
    no_nulos.remove('FEC_INGRESO')

## Correlacion 

data_corr = data[data[col_target].notnull()][no_nulos].corr(method='spearman')

data_corr[col_target].sort_values(ascending=False).apply(lambda _: abs(_))[:58]

#data['ATRASO_PROM_ULT_TRIM_OVER_TRIM_PASADO_MENOR_1'].value_counts(dropna=False)

#select_1.to_excel("VARIABLES_CANDIDATAS_SEGUIR_COVID.xlsx", index=False)

### NUEVAS VARIABLES

El agregado será sobre toda la data

data = all_data.copy()
del all_data

data[data['COD_UNICO'] == str(8443816).zfill(10)].sort_values(by=['PERIODO'])[
    ['target_SF_8_dias_atraso_o_mas_1Mbueno', 'PERIODO', 'MAX_DIAS_ATRASO', 'MAX_DIAS_ATRASO_SF',
     'TIENE_DEUDA_REFINANCIADO_RCC', 'TIENE_DEUDA_REFINANCIADO_IBK', 'TIENE_DEUDA_JUDICIAL_RCC']
]

display(cortes['PERCENT_DELAY_OVER_DEBT_RCC'])

data = add_new_var_x_cut_up(data, colum='ATRASO_PROM_U6M_COLDIRECT_VIG_VENC_SF', cortes=[2,3,4,5,6,7], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='ATRASO_PROM_U3M_COLDIRECT_VIG_VENC_SF', cortes=[2,3,4,5,6,7], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='PERCENT_DELAY_OVER_DEBT_RCC', cortes=[0.01, 0.02, 0.03, 0.04, 0.05, 0.1], impute=0, up=True, soft=False)

display(cortes['MTO_ATRASO_NOLETRAS'])

data = add_new_var_x_cut_up(data, colum='MTO_ATRASO_NOLETRAS', cortes=[20000, 25000, 15000], impute=0, up=True, soft=False)

data['NO_CUENTA_CON_CEF_COMPARABLE'] = data['DESFASE_CEF_MESES_REAL'].apply(lambda _: 1 if pd.isnull(_) else 0)
pd.crosstab(
    data['NO_CUENTA_CON_CEF_COMPARABLE'], data['PERIODO']
)

display(cortes['DESFASE_CEF_MESES_REAL'])

data = add_new_var_x_cut_up(data, colum='DESFASE_CEF_MESES_REAL', cortes=[3,4,5,6,7,8,9,10], impute=0, up=True, soft=False)

display(cortes['MONTO_ATRASO_TOTAL_SF'])

data = add_new_var_x_cut_up(data, colum='MONTO_ATRASO_TOTAL_SF', cortes=[20000, 25000, 35000, 50000, 100000, 160000], impute=0, up=True, soft=False)

display(cortes['DEUDA_IBK_FUENTE_INTERNA'])

data = add_new_var_x_cut_up(data, colum='DEUDA_IBK_FUENTE_INTERNA', cortes=[200000, 100000, 50000, 150000, 250000], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[0]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='ATRASO_PROM_ULT_TRIM_OVER_TRIM_PASADO', cortes=[1, 1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='M_Quartile', cortes=[4], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='M_Quartile', cortes=[3], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='M_Septile', cortes=[4], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='M_Septile', cortes=[5], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='M_Septile', cortes=[6], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='M_Septile', cortes=[7], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='R_Quartile', cortes=[4], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='R_Quartile', cortes=[3], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='R_Septile', cortes=[4], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='R_Septile', cortes=[5], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='R_Septile', cortes=[6], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='R_Septile', cortes=[7], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='F_Quartile', cortes=[4], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='F_Quartile', cortes=[3], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='F_Septile', cortes=[4], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='F_Septile', cortes=[5], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='F_Septile', cortes=[6], impute=0, up=True, soft=False)
data = add_new_var_x_cut_up(data, colum='F_Septile', cortes=[7], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[1]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='MAX_DIAS_ATRASO_SF', cortes=[1, 2,3,4,5,6,7], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='ATRASO_PROM_U12M_COLDIRECT_VIG_VENC_SF', cortes=[5,6,7,8,9, 10], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[2]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_MESES_CON_ATRASO_U6M_COLDIRECT_SF', cortes=[1,2,3,4,5,6], impute=0, up=True, soft=False)

data['impacto_meedio_o_alto'] = data['IMPACTO_COVID_FICHA'].apply(lambda _: 1 if _ in ['alto', 'medio'] else 0)

if creation_mode:
    var = vars_top[3]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASOSF_ULT_RCC', cortes=[1,2,3, 4], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[4]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASO_MAYOR_A_IBKSF_ULT_RCC', cortes=[1,2,3], impute=0, up=True, soft=False)

#display(cortes['SUBSECTOR'])

texto_a_comparar = """
Textil y calzado%,%Restaurante y eventos sociales%,%Bebidas alcohólicas%,%Taller Automotriz%,%Turismo%,%Vehículos%,%Estudio Legal y contable%,%Minerales metálicos%,%Transporte Urbano%,%Hotelería
""".upper()

otro_texto_adicional = """
Abarrotes%,%Agencia de Aduanas y servicios relacionados%,%Artículos de oficina%,%Limpieza%,%Materiales de construcción y ferretería%,%Vehículos, repuestos y accesorios%,%Obras civiles
""".upper()

data['cut_letal_SUBSECTOR'] = data['SUBSECTOR'].apply(lambda _: None if pd.isnull(_) else (1 if _.upper() in texto_a_comparar.upper() else 0))
data['cut_deCuidado_SUBSECTOR'] = data['SUBSECTOR'].apply(
    lambda _: None if pd.isnull(_) else (1 if _.upper() in texto_a_comparar or _.upper() in otro_texto_adicional.upper() else 0))

#pd.crosstab(data['cut_letal_SUBSECTOR'], data['SUBSECTOR'])

if creation_mode:
    var = vars_top[5]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='ENDEUDAMIENTO_PAS_VEN_VAR_CEF_SITUACION', cortes=[0.79999], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[6]
    print(">>>>> ", var)
    display(cortes[var])

display(cortes['NRO_PROD_DEUDA_VIGENTE_SF_ULT_RCC'])

data = add_new_var_x_cut_up(data, colum='NRO_PROD_DEUDA_VIGENTE_SF_ULT_RCC', cortes=[3, 4, 5, 6], impute=0, up=True, soft=False)

from datetime import datetime

data['cambio_accionistas_o_cambio_gerencia'] = data[['CAMBIO_ACCIONISTAS_ANNIO', 'CAMBIO_GERENCIA_GENERAL_ANNIO']].apply(
    lambda _: _[0] == datetime.now().year or _[1] == datetime.now().year, axis=1
).astype(int)

pd.crosstab(
    data['cambio_accionistas_o_cambio_gerencia'], data[col_target].fillna(0)
)

if creation_mode:
    var = vars_top[7]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='COL_DIRECTA_PROMEDIO_MENSUAL_U3M', cortes=[950000], impute=0, up=True, soft=False)

#ROTACION_INVENTARIOS_VAR_DIAS
[col for col in data.columns if 'ROTACION_INVENTARIOS_VAR_DIAS' in col]

#data['ROTACION_INVENTARIOS_VAR_DIAS_CEF_SITUACION_x'] == data['ROTACION_INVENTARIOS_VAR_DIAS_CEF_SITUACION_y']

display(cortes['ROTACION_INVENTARIOS_VAR_DIAS_CEF_SITUACION'])

data = add_new_var_x_cut_up(data, colum='ROTACION_INVENTARIOS_VAR_DIAS_CEF_SITUACION', cortes=[10, 15, 20, 25, 30,35,40,50,60, 80,85,90, 100,150, 270], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[8]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[9]
    print(">>>>> ", var)
    display(cortes[var])

[col for col in data.columns if 'SOBREGIROS' in col.upper()]

data['DEBT_SOBREGIROS__AVANCES_PERCENT_SF'] = data['DEBT_SOBREGIROS__AVANCES_VENCIDO_PERCENT_ULT_RCC']  + data['DEBT_SOBREGIROS__AVANCES_VIGENTE_PERCENT_ULT_RCC']

data = add_new_var_x_cut_up(data, colum='DEBT_SOBREGIROS__AVANCES_PERCENT_SF', cortes=[0.05, 0.1, 0.15, 0.2 , 0.25, 0.3, 0.35, 0.29, 0.28], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[10]
    print(">>>>> ", var)
    display(cortes[var])

[col for col in data.columns if 'MAX_DIAS_ATRASO_SFSF_ULT_RCC' in col.upper()]

del data['MAX_DIAS_ATRASO_SFSF_ULT_RCC_x']
data['MAX_DIAS_ATRASO_SFSF_ULT_RCC'] = data['MAX_DIAS_ATRASO_SFSF_ULT_RCC_y'].copy()
del data['MAX_DIAS_ATRASO_SFSF_ULT_RCC_y']

data = add_new_var_x_cut_up(data, colum='MAX_DIAS_ATRASO_SFSF_ULT_RCC', cortes=[1, 2, 3, 4], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[11]
    print(">>>>> ", var)
    display(cortes[var])


#data = add_new_var_x_cut_up(data, colum='CEF_ROTACION_INVENTARIOS', cortes=[200], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[12]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[13]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='MAX_DIAS_ATRASO_SFSF_ULT_RCC', cortes=[1, 2, 3, 4], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[14]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_DOCS_IMPAG_U1M', cortes=[2,3,4,5,6], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='DTOTAL_S_VENTAS_CEF_SITUACION', cortes=[0.5, 0.6, 0.7,0.8,0.9, 1, 1.1, 1.2,1.3], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[15]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[16]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[17]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='CAPITAL_TRABAJO_CEF_SITUACION', cortes=[0], impute=0, up=False, soft=False)

data = add_new_var_x_cut_up(data, colum='MAX_DIAS_ATRASO_SF', cortes=[1,2,3,4,5,6,7], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[18]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[19]
    print(">>>>> ", var)
    display(cortes[var])

[col for col in data.columns if 'ENDEUDAMIENTO_PAS_VEN_VAR' in col.upper()]

display(cortes['ENDEUDAMIENTO_PAS_VEN_VAR_CEF_SITUACION'])

data = add_new_var_x_cut_up(data, colum='ENDEUDAMIENTO_PAS_VEN_VAR_CEF_SITUACION', cortes=[0.19, 0.2, 0.29, 0.3, 0.39, 0.4, 0.49, 0.5, 1, 1.1, 1.2, 0.6, 0.7, 0.8, 0.9], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[20]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[21]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[22]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[23]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[24]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[25]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='BUCKET', cortes=[3, 4], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[26]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[27]
    print(">>>>> ", var)
    display(cortes[var])

if creation_mode:
    var = vars_top[28]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='MAX_DIAS_ATRASO_SFSF_ULT_RCC', cortes=[2, 5], impute=0, up=True, soft=False)

if creation_mode:
    var = vars_top[29]
    print(">>>>> ", var)
    display(cortes[var])

data = add_new_var_x_cut_up(data, colum='VENTAS_NETAS_VAR_CEF_SITUACION', cortes=[0, -0.1, -0.2, -0.3, -0.4, -0.5, -0.6], impute=0, up=False, soft=True)

if creation_mode:
    var = vars_top[30]
    print(">>>>> ", var)
    display(cortes[var])

var = vars_top[31]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='MTO_DEUDA_FULL_RCC_IBK', cortes=[26500000.0], impute=0, up=True, soft=False)

var = vars_top[32]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[33]
print(">>>>> ", var)
display(cortes[var])

[col for col in data.columns if 'ROTACION_CTASXCOBRAR_VAR' in col.upper()]

#display(cortes['ROTACION_CTASXCOBRAR_VAR_DIAS_CEF_SITUACION'])

display(cortes['ROTACION_CTASXCOBRAR_VAR_DIAS_CEF_SITUACION'])

data = add_new_var_x_cut_up(data, colum='ROTACION_CTASXCOBRAR_VAR_DIAS_CEF_SITUACION', cortes=[5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 60, 55, 65, 70, 75, 80, 85, 90, 100, 110, 120], impute=0, up=True, soft=False)

display(cortes['ROTACION_CTASXCOBRAR_VAR_CEF_SITUACION'])

data = add_new_var_x_cut_up(data, colum='ROTACION_CTASXCOBRAR_VAR_CEF_SITUACION', cortes=[0.2, 0.29, 0.3, 0.39, 0.49, 0.59, 0.6, 0.69, 0.79, 0.8, 0.89, 0.9, 1, 1.1, 1.2, 1.19], impute=0, up=True, soft=False)

var = vars_top[34]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='PERCENT_GRUPO_ECON_VPC_SEGUIR', cortes=[0.1], impute=0, up=True, soft=False)

var = vars_top[35]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='IMPORTE_TRIB_PROMEDIO_U9M', cortes=[20000], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASOSF_ULT_RCC', cortes=[1,2,3,4], impute=0, up=True, soft=False)

var = vars_top[36]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[37]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASO_MAYOR_A_IBKSF_ULT_RCC', cortes=[1,2,3,4], impute=0, up=True, soft=False)

var = vars_top[38]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='DTOTAL_S_EBITDA_VAR_CEF_SITUACION', cortes=[1, 1.5, 2, 2.5, 3, 3.5], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='TIPOS_CASH_IN_ACTIVOS_U3M', cortes=[5], impute=0, up=True, soft=False)

var = vars_top[39]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_DOCS_IMPAG_U1M', cortes=[3,4,5], impute=0, up=True, soft=False)

var = vars_top[40]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[41]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_FEVE_DURO_GRUPO_EECON', cortes=[1,2,3], impute=0, up=True, soft=False)

var = vars_top[42]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[43]
print(">>>>> ", var)
display(cortes[var])

#del data['_60_cut_up_CEF_ROTACION_CTASXPAGAR_VAR']

#data = add_new_var_x_cut_up(data, colum='CEF_ROTACION_CTASXPAGAR_VAR', cortes=[10, 30, 60], impute=0, up=True, soft=False)

var = vars_top[44]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='BUCKET', cortes=[2,3,4], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='DTOTAL_S_EBITDA_CEF_SITUACION', cortes=[1, 2 , 3, 4, 5, 6], impute=0, up=True, soft=False)

#data = add_new_var_x_cut_up(data, colum='DEBT_PORM_u6m_anio_act', cortes=[12500000.0], impute=0, up=True, soft=False)

var = vars_top[45]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='IMPORTE_IMPAG_PROMEDIO_U9M', cortes=[1150], impute=0, up=True, soft=False)

var = vars_top[46]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASO_MAYOR_A_IBKSF_ULT_RCC', cortes=[2], impute=0, up=True, soft=False)

var = vars_top[47]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASOSF_ULT_RCC', cortes=[2, 3], impute=0, up=True, soft=False)

var = vars_top[48]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[49]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[50]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[51]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[52]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='MTO_DEUDA_TOTAL_COLDIRECT_SIN_REATCIVA_SF', cortes=[13500000.0], impute=0, up=True, soft=False)

var = vars_top[53]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[54]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_MESES_ANTERIORES_SIN_IMPORTE_TRIB', cortes=[1], impute=0, up=False, soft=False)



for column in [col for col in data.columns if 'NRO_MESES_ANTERIORES_CON' in col.upper()]:
    print(column)
    data = add_new_var_x_cut_up(data, colum=column, cortes=[1,2,3,5,7,9,10], impute=0, up=True, soft=False)

var = vars_top[55]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASO_MAYOR_A_IBKSF_ULT_RCC', cortes=[1, 2, 3], impute=0, up=True, soft=False)

var = vars_top[56]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='DTOTAL_S_VENTAS_VAR_CEF_SITUACION', cortes=[0.39, 0.49, 0.59, 0.69, 0.79, 0.89, 1, 1.1, 1.2, 1.2, 1.3, 1.4, 1.5, 1.59, 1.6], impute=0, up=True, soft=False)

var = vars_top[57]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='DIFF_PROM_CANT_TRX_U6M_U12M_CASHIN', cortes=[-5, -3, -10, -20, -15], impute=0, up=False, soft=True)

display(cortes['DEBT_VENCIDO_PERCENT_ULT_RCC'])

data = add_new_var_x_cut_up(data, colum='DEBT_VENCIDO_PERCENT_ULT_RCC', cortes=[0, 0.001], impute=0, up=True, soft=False)

var = vars_top[58]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='RSOL_18_SOLVENCIA_CEF_SITUACION', cortes=[0.19, 0.29, 0.39, 0.49, 0.59, 0.69, 0.7, 0.79, 0.8], impute=0, up=True, soft=False)

data['DIFF_PROM_CANT_TRX_U6M_U12M_CASHIN'].min()

var = vars_top[59]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='DIFF_PROM_CANT_TRX_U6M_U12M_CASHIN', cortes=[-19, -12, 0], impute=0, up=False, soft=True)

var = vars_top[60]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASOSF_ULT_RCC', cortes=[1, 2], impute=0, up=True, soft=False)

var = vars_top[61]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[62]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[63]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[64]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[65]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[66]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[67]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[68]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[69]
print(">>>>> ", var)
display(cortes[var])

#data = add_new_var_x_cut_up(data, colum='TENDENCIA_PERCENT_CLASIF_NO_NORMAL_GRUPO_ECON_U3M', cortes=[0.01], impute=0, up=True, soft=False)

var = vars_top[70]
print(">>>>> ", var)
display(cortes[var])

display(cortes['COVER_GAR_AUTO_HIPO_ULT_RCC'])

data = add_new_var_x_cut_up(data, colum='COVER_GAR_AUTO_HIPO_ULT_RCC', cortes=[0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7], impute=0, up=False, soft=True)

var = vars_top[71]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='ROTACION_INVENTARIOS_VAR_CEF_SITUACION', cortes=[0.29, 0.39, 0.39, 0.49, 0.59, 0.69, 0.79, 0.8, 0.89, 1, 1.1, 1.2, 1.3, 1.39, 1.4, 1.5], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_CON_DIAS_ATRASO_MAYOR_4SF_ULT_RCC', cortes=[1, 2, 3], impute=0, up=True, soft=False)

var = vars_top[72]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[73]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[74]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[75]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[76]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[77]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[78]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[79]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[80]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[81]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[82]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[83]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[84]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[85]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[86]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[87]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[88]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[89]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='APLANCAMIENTO_PATRIMONIAL_VAR_CEF_SITUACION', cortes=[0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 1.1, 1.2, 1.3, 1.4, 1.5, 1.6], impute=0, up=True, soft=False)

var = vars_top[90]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='ROTACION_CTASXPAGAR_VAR_DIAS_CEF_SITUACION', cortes=[5, 10, 20, 30, 60, 90, 120, 50, 40, 35, 25, 65, 100, 150, 200, 15], impute=0, up=True, soft=False)

var = vars_top[91]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='ROTACION_CTASXPAGAR_VAR_CEF_SITUACION', cortes=[0.1, 0.2, 0.29, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 1.1, 1.2], impute=0, up=True, soft=False)

var = vars_top[92]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[93]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[94]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[95]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[96]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[97]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[98]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[99]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='COSTO_VENTAS_VAR_CEF_SITUACION', cortes=[0, -0.1, -0.19, -0.39, -0.4], impute=0, up=False, soft=True)

data = add_new_var_x_cut_up(data, colum='COSTO_VENTAS_VAR_CEF_SITUACION', cortes=[0, 0.1, 0.2, 0.3], impute=0, up=True, soft=False)

var = vars_top[100]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='RMET_9_METODIZADO_CEF_SITUACION', cortes=[0, 1, 1.5, 2, 2.5, 3, 3.5, 4], impute=0, up=True, soft=False)

var = vars_top[101]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[102]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[103]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[104]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='ROTACION_INVENTARIOS_CEF_SITUACION', cortes=[20, 30, 60, 100, 120, 150, 180, 200, 250, 300, 350, 205, 210, 190, 270], impute=0, up=True, soft=False)

var = vars_top[105]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='SOW_OTROSBCOS_MIN_historia', cortes=[0.6, 0.7, 0.8, 0.9, 0.65, 0.75, 0.85, 0.95], impute=0, up=True, soft=False)

var = vars_top[106]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[107]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[108]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='TRX_TIPOS_CASH_OUT_CAIDA_10PERCENT_MES_ANT', cortes=[2, 3], impute=0, up=True, soft=False)

var = vars_top[109]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[110]
print(">>>>> ", var)
display(cortes[var])


display(cortes['RECENCY_historia'])

data = add_new_var_x_cut_up(data, colum='RECENCY_historia', cortes=[32, 30, 35, 40, 25, 50, 60, 55, 70, 80, 90], impute=0, up=True, soft=False)

var = vars_top[111]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_PRIN_DEUDASF_ULT_RCC', cortes=[2,3,4,5], impute=0, up=True, soft=False)

var = vars_top[112]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='ROTACION_CTASXCOBRAR_VAR_DIAS_CEF_SITUACION', cortes=[15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 80, 90], impute=0, up=True, soft=False)

var = vars_top[113]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='DTOTALCAJA_S_EBITDA_VAR_CEF_SITUACION', cortes=[1,2,3,1.5,2.5,3.5], impute=0, up=True, soft=False)

var = vars_top[114]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_ENTIDADES_FULL_RCC_SF', cortes=[3,4,5,6], impute=0, up=True, soft=False)

data = add_new_var_x_cut_up(data, colum='ROTACION_INVENTARIOS_VAR_DIAS_CEF_SITUACION', cortes=[10,20,30,40,50,60,70,80,90, 15,25,35,45,55,65,75,85,95], impute=0, up=True, soft=False)

var = vars_top[115]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[116]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='RIESGO', cortes=[80000.0, 90000.0, 100000.0, 400000.0, 500000.0, 600000.0, 700000.0], impute=0, up=True, soft=False)

var = vars_top[117]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[118]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[119]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[120]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[121]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[122]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[123]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='minimo_percent_sow_others_u12m', cortes=[0.6, 0.7, 0.8, 0.9], impute=0, up=True, soft=False)

var = vars_top[124]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[125]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[126]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[127]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[128]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_MESES_CON_ATRASO_U6M_COLDIRECT_SF', cortes=[1, 2, 3, 4], impute=0, up=True, soft=False)

var = vars_top[129]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[130]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_MESES_CON_ATRASO_U6M_COLDIRECT_SF', cortes=[3, 4], impute=0, up=True, soft=False)

var = vars_top[131]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[132]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[133]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='NRO_CLASIF_NO_NORMAL_GRUPO_EECON', cortes=[1, 2, 3, 4, 5, 6, 7, 8], impute=0, up=True, soft=False)

var = vars_top[134]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[135]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[136]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[137]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[138]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[139]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[139]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[140]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[141]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[142]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[143]
print(">>>>> ", var)
display(cortes[var])

data = add_new_var_x_cut_up(data, colum='maximo_percent_ibk_u12m', cortes=[0.2, 0.3], impute=0, up=False, soft=True)

data = add_new_var_x_cut_up(data, colum='VARACION_NUM_BANCOS_FULL_RCCC_SF', cortes=[1, 2, 3], impute=0, up=True, soft=False)

var = vars_top[144]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[145]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[146]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[147]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[148]
print(">>>>> ", var)
display(cortes[var])

var = vars_top[149]
print(">>>>> ", var)
display(cortes[var])

display(cortes['PERCENT_GRUPO_ECON_VPC_NO_CLASIF_NORMAL'])

data = add_new_var_x_cut_up(data, colum='PERCENT_GRUPO_ECON_VPC_NO_CLASIF_NORMAL', cortes=[0.1, 0.2, 0.3, 0.4], impute=0, up=True, soft=False)

#[col for col in data.columns if 'ECON' in col.upper()]

data = add_new_var_x_cut_up(data, colum='NRO_CLASIF_NO_NORMAL_GRUPO_EECON', cortes=[1, 2, 3], impute=0, up=True, soft=False)


data = add_new_var_x_cut_up(data, colum='NRO_MESES_CON_ATRASO_4diasomas_U6M_COLDIRECT_SF', cortes=[1, 2, 3, 4], impute=0, up=True, soft=False)


data = add_new_var_x_cut_up(data, colum='NRO_BANCOS_CON_MAYOR_SOW_ULT_RCC', cortes=[1, 2, 3], impute=0, up=True, soft=False)

## OTRAS VARIABLES

for var in [col for col in data.columns if 'full' in col.lower()]:
    try:
        print(">"*20, ' ', var)
        display(cortes[var])
    except Exception as e:
        print(str(e))

#data['VARIAICION_NRO_BANCOS_CIERRE_2019'].value_counts()

data['fusion_atraso_grantia'] = data[['TIENE_MAS_95PERCENT_GAR_TOTAL_EN_UN_SOLO_BANCO_PRIN_NO_IBK', '_1_cut_up_NUM_BANCOS_PRIN_CON_DIAS_ATRASO_MAYOR_A_IBKSF_ULT_RCC']].apply(
    lambda _: 1 if _[0] == 1 or _[1] == 1 else 0, axis=1
)

for column in [col for col in data.columns if 'NRO_BANCO' in col.upper() and 'and' not in col and 'VARIAICION' not in col]:
    data = add_new_var_x_cut_up(data, colum=column, cortes=[1], impute=0, up=True, soft=False)

for column in [col for col in data.columns if 'TENDENCIA' in col.upper() and 'and' not in col]:
    try:
        data = add_new_var_x_cut_up(data, colum=column, cortes=[0], impute=0, up=True, soft=True)
    except:
        print(">>ERROR: ", column)

#[col for col in data.columns if 'PROTESTO' in col.upper()]
for var in [col for col in data.columns if 'protesto' in col.lower()]:
    try:
        print(">"*20, ' ', var)
        display(cortes[var])
    except Exception as e:
        print(str(e))

#data['NUM_ENTIDADES_FULL_RCC_SF']


for var in [col for col in data.columns if 'NUM_ENTIDADES_FULL_RCC_SF' in col.upper()]:
    try:
        print(">"*20, ' ', var)
        display(cortes[var])
    except Exception as e:
        print(str(e))

#for var in [col for col in data.columns if 'riesgo' in col.lower()]:
#    try:
#        print(">"*20, ' ', var)
#        display(cortes[var])
#    except Exception as e:
#        print(str(e))

### IMPUTANDO VARIABLES DICOTOMIAS

data[data['COD_UNICO'] == str(8443816).zfill(10)].sort_values(by=['PERIODO'])[
    ['target_SF_8_dias_atraso_o_mas_1Mbueno', 'PERIODO', 'MAX_DIAS_ATRASO', 'MAX_DIAS_ATRASO_SF',
     'TIENE_DEUDA_REFINANCIADO_RCC', 'TIENE_DEUDA_REFINANCIADO_IBK', 'TIENE_DEUDA_JUDICIAL_RCC']
]

for termino in ['ATRASO', 'CATEGORIA', 'TIENE', 'CUT', 'MENOR', 'MAYOR', 'SUBIDA',
                'NUM', 'INT', 'CAIDA', 'NRO', 'COVID', 'BANCO']:
    
    for col in [_ for _ in data.columns if termino in _.upper() and 'TARGET' not in _.upper()]:
        data[col] = data[col].fillna(0)
    


data[data['COD_UNICO'] == str(8443816).zfill(10)].sort_values(by=['PERIODO'])[
    ['target_SF_8_dias_atraso_o_mas_1Mbueno', 'PERIODO', 'MAX_DIAS_ATRASO', 'MAX_DIAS_ATRASO_SF',
     'TIENE_DEUDA_REFINANCIADO_RCC', 'TIENE_DEUDA_REFINANCIADO_IBK', 'TIENE_DEUDA_JUDICIAL_RCC']
]

## SAVE

#[col for col in data.columns if 'categoria' in col.lower()]

data.columns = [col.replace(' ', '_') for col in data.columns]

data = data[[col for col in data.columns if '1_and_1' not in col]]
print(data.shape)

len([col for col in data.columns if 'fusion_atraso_grantia'  in col])

'TIENE_AUMENTO_RIESGO_U7M' in data.columns, '_1_cut_up_NUM_BANCOS_NOIBK_DEBT_MAYOR_DE_U3M' in data.columns, 'NRO_BANCOS_CON_MAYOR_SOW_ULT_RCC' in data.columns

data.shape

data[[col for col in data.columns if 'and' not in col]].to_csv("AVANCE_6_MEDIANA.csv", index=False)

pd.crosstab(
    data[data['FUNNEL_4'] == 1]['ZONAL'], data[data['FUNNEL_4'] == 1]['PERIODO']
)

pd.crosstab(
    data[data['FUNNEL_ALERTA'] == 1]['ZONAL'], data[data['FUNNEL_ALERTA'] == 1]['PERIODO']
)

data.shape
