==> courses - workshops/ALERTAS EMPRESA/3.0_CONSOLIDADO.ipynb <==
# Nativos
import random as rn
import os
import sys
import gc

#calculo
import numpy as np
import pandas as pd
import scipy

#grafico
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
%matplotlib inline
sns.set(style="whitegrid")

#warning ignore future
import warnings
# warnings.simplefilter(action='ignore', category=FutureWarning)
warnings.filterwarnings("ignore")
gc.collect()

SEED = 29082013
os.environ['PYTHONHASHSEED'] = str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

tipos_llave = {
    'COD_UNICO': str,
    'PERIODO': str
}
llave = list(tipos_llave.keys())
llave

## MEDIANAS

mediana = pd.read_csv('AVANCE_6_MEDIANA.csv', dtype=tipos_llave)
print(mediana.shape)
mediana.head()

mediana.shape

cols_habilitadas = {col for col in mediana.columns if '1_and_1' not in col.lower() and 'listado' not in col.lower()}
print(len(cols_habilitadas))

base = mediana[cols_habilitadas].copy()
del mediana

base.shape

base[base['COD_UNICO'] == str(8443816).zfill(10)].sort_values(by=['PERIODO'])[
    ['target_SF_8_dias_atraso_o_mas_1Mbueno', 'PERIODO', 'MAX_DIAS_ATRASO', 'MAX_DIAS_ATRASO_SF',
     'TIENE_DEUDA_REFINANCIADO_RCC', 'TIENE_DEUDA_REFINANCIADO_IBK', 'TIENE_DEUDA_JUDICIAL_RCC']
]

## GRAN EMPRESA

ge = pd.read_csv('AVANCE_6_GRAN_EMPRESA.csv', dtype=tipos_llave)
print(ge.shape)
ge.head(2)

cols_new = {col for col in ge.columns if '1_and_1' not in col.lower() and 'listado' not in col.lower()} - set(base.columns)
print(len(cols_new))

print(base.shape)
base = base.merge(
    ge[list(cols_new) + llave], how='left', on=llave
)
print(base.shape)

del ge

len(list(base.columns)), len(set(base.columns))

## S1 PROVINCIA

s1 = pd.read_csv('AVANCE_6_S1_PROVINCIA.csv', dtype=tipos_llave)
print(s1.shape)
s1.head(2)

cols_new = {col for col in s1.columns if '1_and_1' not in col.lower() and 'listado' not in col.lower()} - set(base.columns)
print(len(cols_new))

print(base.shape)
base = base.merge(
    s1[list(cols_new) + llave], how='left', on=llave
)
base.shape

del s1

## S2 PROVINCIA

s2 = pd.read_csv('AVANCE_6_S2_PROVINCIA.csv', dtype=tipos_llave)
s2.head(2)

cols_new = {col for col in s2.columns if '1_and_1' not in col.lower()} - set(base.columns)
print(len(cols_new))

print(base.shape)
base = base.merge(
    s2[list(cols_new) + llave], how='left', on=llave
)
base.shape

del s2

base[base['COD_UNICO'] == str(8443816).zfill(10)].sort_values(by=['PERIODO'])[
    ['target_SF_8_dias_atraso_o_mas_1Mbueno', 'PERIODO', 'MAX_DIAS_ATRASO', 'MAX_DIAS_ATRASO_SF',
     'TIENE_DEUDA_REFINANCIADO_RCC', 'TIENE_DEUDA_REFINANCIADO_IBK', 'TIENE_DEUDA_JUDICIAL_RCC']
]

#### SAVE

base.to_csv('AVANCE_6.csv', index=False)

#base.to_excel('AVANCE_6.xlsx', index=False)

len(list(base.columns)), len(set(base.columns))

del base

#(2479, 2479)
