==> courses - workshops/ALERTAS EMPRESA/3.4_S2_PROVINCIA_SELECCION_ALERTAS.ipynb <==
# Nativos
import random as rn
import os
import sys
import gc
import time

#calculo
import numpy as np
import pandas as pd
import scipy

#grafico
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
%matplotlib inline
sns.set(style="whitegrid")

#warning ignore future
import warnings
# warnings.simplefilter(action='ignore', category=FutureWarning)
warnings.filterwarnings("ignore")
gc.collect()

SEED = int(time.time() / 1000)
os.environ['PYTHONHASHSEED'] = str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

from utils_alertas import *

data = pd.read_csv('AVANCE_6.csv', dtype={
    'COD_UNICO': str,
    'PERIODO': str
})
data.head()

'COD_UNICO' in data.columns, 'PERIODO' in data.columns

print(data.shape)
data = filtro_banca_s2provincia(data)
print(data.shape)

pd.crosstab(
    data['PERIODO'], data['ZONAL']
)

### CREACION DE COLUMNAS A MEZCLAR

cols_combine = [col for col in data.columns if set(data[col].unique()) == {1, 0} if '1_and_1' not in col]
print(len(cols_combine))

cols_combine = \
    [col for col in cols_combine if 'SCOTIA' not in col.upper() \
                                 and 'BCP' not in col.upper() \
                                 and 'BBVA' not in col.upper() \
                                 and 'BIF' not in col.upper() \
                                 and '%' not in col.upper() \
                                 and 'COVID' not in col.upper() \
                                 and 'REFIN' not in col.upper() 
    ]
print(len(cols_combine))

for col in ['cambio_seguir_covid', 'TIENE_DIAS_DE_ATRASO_IBK_ESTE_MES'] + [col for col in data.columns if 'target' in col.lower()]:
    if col in cols_combine:
        cols_combine.remove(col)

len(cols_combine)

### GENERACION DE ALERTAS

existentes = pd.read_csv('COMBINACIONES_S2_PROVINCIA.csv', sep=',').sort_values(by=['SCORE'], ascending=False)
existentes.head(50)

#existentes.to_csv('COMBINACIONES_S2_PROVINCIA.csv', index=False)

try:
    existentes['union'] = existentes[['VAR_1', 'VAR_2', 'VAR_3']].apply(
        lambda _: '1_and_1_{}___{}'.format(_[0], _[1]) if pd.isnull(_[2]) else '1_and_1_and_1_{}___{}____{}'.format(_[0], _[1], _[2]), 
        axis=1
    )
    listado = list(existentes['union'].unique())
    len(listado)
except:
    listado = []
    
len(listado)

import random as rn
import itertools

col_target = 'target_FINAL_fusion_atraso_feve_2Mbueno'
display(pd.crosstab(
    data[col_target].fillna(-1), data['PERIODO']
))

aucc = {}
struct = []

data = data[data[col_target].notnull()][cols_combine + ['COD_UNICO', col_target, 'PERIODO']]
data.shape

#pd.DataFrame([], columns=['VAR_1', 'VAR_2', 'VAR_3', 'CASOS', 'CORRECTOS', 'SCORE']).to_csv('COMBINACIONES_MEDIANA_LIMA.csv', index=False)
#pd.DataFrame([], columns=['VAR_1', 'VAR_2', 'VAR_3', 'CASOS', 'CORRECTOS', 'SCORE']).to_csv('COMBINACIONES_GRAN_EMPRESA_LIMA.csv', index=False)
#pd.DataFrame([], columns=['VAR_1', 'VAR_2', 'VAR_3', 'CASOS', 'CORRECTOS', 'SCORE']).to_csv('COMBINACIONES_S1_PROVINCIA.csv', index=False)
#pd.DataFrame([], columns=['VAR_1', 'VAR_2', 'VAR_3', 'CASOS', 'CORRECTOS', 'SCORE']).to_csv('COMBINACIONES_S2_PROVINCIA.csv', index=False)

#['TIENE_MENOS_5PERCENT_GARTOTAL_IBK', 'MORE_0.11_PERCENT_DELAY_OVER_TOTAL_DEBT_CC', None, 12, 9, 0.75]

SEED = int(time.time() / 1000)
os.environ['PYTHONHASHSEED'] = str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

rn.shuffle(cols_combine)

over_interes = [col for col in data.columns if ('OVER_TRIM_PASADO_MAYOR' in col.upper() or 'CIERRE' in col.upper()) and '1_and_1' not in col]
over_interes

from datetime import datetime
tiempo_inicio = time.time()
nn = len(struct)

while True:
    x = rn.choice(over_interes)
    y = rn.choice(cols_combine)
    z = rn.choice(cols_combine)
    
    col_evaluate = '1_and_1_and_1_{}___{}____{}'.format(x, y, z)
    
    if col_evaluate not in listado and x != y and x!= z and y!= z:
        listado.append(col_evaluate)
        
        lambda_basico = lambda _:  _[0] >= 1 and _[1] >= 1 and _[2] >= 1
        data[col_evaluate] = data[[x, y, z]].apply(lambda_basico, axis=1).astype(int)
        #data_play_1[col_evaluate] = data_play_1[[x, y, z]].apply(lambda_basico, axis=1).astype(int)
        
        
        coincidencias = len(set(data[data[col_evaluate] == 1]['COD_UNICO'].unique()))
        acertados = len(set(data[(data[col_evaluate] == 1) & (data[col_target] == 1)]['COD_UNICO'].unique()))
        
        if acertados > 6 and acertados/coincidencias > 0.35:
            aucc[col_evaluate] = acertados/coincidencias
            struct.append(
                [x, y, z, coincidencias, acertados, acertados/coincidencias]
            )
            tiempo_inicio = time.time()
            nn = len(struct)
            #print(struct[-1])
            
        ### Generado pares de 2 
        for a, b in list(itertools.combinations([x, y, z], 2)):
            
            col_evaluate = '1_and_1_{}___{}'.format(a, b)
            if col_evaluate not in listado:
                lambda_basico = lambda _:  _[0] >= 1 and _[1] >= 1
                data[col_evaluate] = data[[a, b]].apply(lambda_basico, axis=1).astype(int)
                #data_play_1[col_evaluate] = data_play_1[[a, b]].apply(lambda_basico, axis=1).astype(int)
                
                coincidencias = len(set(data[data[col_evaluate] == 1]['COD_UNICO'].unique()))
                acertados = len(set(data[(data[col_evaluate] == 1) & (data[col_target] == 1)]['COD_UNICO'].unique()))
                
                if acertados > 6 and acertados/coincidencias > 0.35:
                    aucc[col_evaluate] = acertados/coincidencias
                    struct.append(
                        [a, b, None, coincidencias, acertados, acertados/coincidencias]
                    )
                    tiempo_inicio = time.time()
                    nn = len(struct)
                    #print(struct[-1])
        
        if len(struct) >= 5:
            pd.DataFrame(struct).to_csv('COMBINACIONES_S2_PROVINCIA.csv', header=False, index=False, mode='a')
            print("/"*25, datetime.now())
            struct = []
            
        if time.time() - tiempo_inicio >= 3600 and nn == len(struct):
            print("Mas de 1 hora sin nueva combinacion de alerta")
            break
