==> courses - workshops/ALERTAS EMPRESA/6.2 MACRO.ipynb <==
import glob
import pandas as pd
relacion = []

valores = pd.read_csv('_FORMA_VALOR_ALERTA.csv', dtype={
    'COD_UNICO': str,
    'PERIODO': str
})

print(valores.shape)
valores.head()

valores[valores['COD_UNICO'] == '0011027012']

valores[valores['COD_UNICO'] == '0015506613']

pd.crosstab(
    valores['CODIGO_ALERTA'], valores['PERIODO']
).sort_index()

## ULTIMO PERIODO

ult_periodo = valores['PERIODO'].max()
ult_periodo

ultimos_valores = valores[valores['PERIODO'] ==  ult_periodo]
print(ultimos_valores.shape)
ultimos_valores.head()

ultimos_valores.tail()

ultimos_valores.shape

#ultimos_valores[ultimos_valores['COD_UNICO'] == '0006065262']

%%time

from sqlalchemy import create_engine
engine = create_engine(r"mssql+pymssql://ibetlmarket:m@rk3t2o15@10.11.12.90\BDT/MARKET")
connection_x = engine.connect()

try:
    ultimos_valores.to_sql(
        name='H_ALERTAS_CLIENTE_ALERTA',
        schema='STG',
        con=connection_x, 
        index=False, 
        if_exists='replace'
    )
    connection_x.close()
    correct = 1
except Exception as e:
    print("ERROR: ", str(e))
    correct = 0

print(correct)

### HISTÓRICO

%%time

from sqlalchemy import create_engine
engine = create_engine(r"mssql+pymssql://ibetlmarket:m@rk3t2o15@10.11.12.90\BDT/MARKET")
connection_x = engine.connect()

try:
    valores.to_sql(
        name='H_ALERTAS_CLIENTE_ALERTA_HIST',
        schema='STG',
        con=connection_x, 
        index=False, 
        if_exists='replace'
    )
    connection_x.close()
    correct = 1
except Exception as e:
    print("ERROR: ", str(e))
    correct = 0

print(correct)
