==> courses - workshops/ALERTAS EMPRESA/6.3 CHANGE_ROLES_GRUPO.ipynb <==
import glob
import time
import gc

#nativos
import os
import unicodedata
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta

#terceros
import pandas as pd
import numpy as np
import pymssql
import missingno as msnolistadconoce
from collections import Counter

#conneciom MARKET
connection_market = pymssql.connect(
    host='10.11.12.90\BDT',
    user='ibetlmarket',
    port=1433,
    password='m@rk3t2o15',
    database='MARKET'
)

import time
time.sleep(1)
gc.collect()

sql_rol = """
     SELECT  * FROM STG.H_ALERTAS_GRUPO_ROL_EQUIPO 
"""
rol_equipo = pd.read_sql(
   sql_rol , connection_market
)

print(rol_equipo.shape)
rol_equipo.head(2)

rol_equipo.to_excel('GRUPO_ROL_EQUIPO.xlsx', index=False)

sql_colaborador = """
     SELECT  * FROM STG.H_ALERTAS_COLABORADOR_EQUIPO 
"""
rol_colaborador = pd.read_sql(
   sql_colaborador , connection_market
)

print(rol_colaborador.shape)
rol_colaborador.head(2)

rol_colaborador.to_excel('COLABORADOR_EQUIPO.xlsx', index=False)

connection_market.close()

raise Exception("VICTOR DICE: AGUANTA TU COCHE .... valida en el excel y luego corre lo de abajo .... pdta: El excel ya se creó porseaca")

# UP

### GRUPO ROL EQUIPO

grupo_rol_equipo = pd.read_excel('GRUPO_ROL_EQUIPO.xlsx', engine='openpyxl')
grupo_rol_equipo.head()

grupo_rol_equipo.shape

%%time
# mssql+pymssql
from sqlalchemy import create_engine
engine = create_engine(r"mssql+pymssql://ibetlmarket:m@rk3t2o15@10.11.12.90\BDT/MARKET")
connection_x = engine.connect()

try:
    grupo_rol_equipo.to_sql(
        name='H_ALERTAS_GRUPO_ROL_EQUIPO',
        schema='STG',
        con=connection_x, 
        index=False, 
        if_exists='replace'
    )
    connection_x.close()
    correct = 1

except Exception as e:
    print("ERROR: ", str(e))
    correct = 0

print(correct)

import pymssql
print("Inciando actualizacion de tblas ODS")
connection = pymssql.connect(
    host='10.11.12.90\BDT',
    user='ibetlmarket',
    port=1433,
    password='m@rk3t2o15',
    database='MARKET'
)

with connection.cursor() as cursor:
    try:
        print("/"*50)
        alter_text_files = """
        ALTER TABLE STG.H_ALERTAS_GRUPO_ROL_EQUIPO ALTER COLUMN GRUPO VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_GRUPO_ROL_EQUIPO ALTER COLUMN ROL VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_GRUPO_ROL_EQUIPO ALTER COLUMN EQUIPO VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_GRUPO_ROL_EQUIPO ALTER COLUMN ZONAL VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_GRUPO_ROL_EQUIPO ALTER COLUMN SEGMENTO VARCHAR(255);
        """
        cursor.execute(alter_text_files)
    except Exception as e:
        print("ERROR: ", str(e))
        
try:
    connection.commit()
    connection.close()
    print("Conexion database cerrada")
except Exception as e:
    print("Error:  ", str(e))

### COLABORADOR EQUIPO

colaborador_equipo = pd.read_excel('COLABORADOR_EQUIPO.xlsx', engine='openpyxl')
colaborador_equipo.head()

colaborador_equipo.shape

%%time
# mssql+pymssql
from sqlalchemy import create_engine
engine = create_engine(r"mssql+pymssql://ibetlmarket:m@rk3t2o15@10.11.12.90\BDT/MARKET")
connection_x = engine.connect()

try:
    colaborador_equipo.to_sql(
        name='H_ALERTAS_COLABORADOR_EQUIPO',
        schema='STG',
        con=connection_x, 
        index=False, 
        if_exists='replace'
    )
    connection_x.close()
    correct = 1

except Exception as e:
    print("ERROR: ", str(e))
    correct = 0

print(correct)

import pymssql
print("Inciando actualizacion de tblas ODS")
connection = pymssql.connect(
    host='10.11.12.90\BDT',
    user='ibetlmarket',
    port=1433,
    password='m@rk3t2o15',
    database='MARKET'
)

with connection.cursor() as cursor:
    try:
        print("/"*50)
        alter_text_files = """
        ALTER TABLE STG.H_ALERTAS_COLABORADOR_EQUIPO ALTER COLUMN REGISTRO VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_COLABORADOR_EQUIPO ALTER COLUMN NOMBRE VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_COLABORADOR_EQUIPO ALTER COLUMN PUESTO VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_COLABORADOR_EQUIPO ALTER COLUMN EQUIPO VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_COLABORADOR_EQUIPO ALTER COLUMN ZONAL VARCHAR(255);
        ALTER TABLE STG.H_ALERTAS_COLABORADOR_EQUIPO ALTER COLUMN SEGMENTO VARCHAR(255);
        """
        cursor.execute(alter_text_files)
    except Exception as e:
        print("ERROR: ", str(e))
        
try:
    connection.commit()
    connection.close()
    print("Conexion database cerrada")
except Exception as e:
    print("Error:  ", str(e))
