==> courses - workshops/ALERTAS EMPRESA/6.5_Seguimiento_gestion.ipynb <==
import glob
import time
import gc

#nativos
import os
import unicodedata
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta

#terceros
import pandas as pd
import numpy as np
import pymssql
import missingno as msnolistadconoce
from collections import Counter

#conneciom MARKET
connection_market = pymssql.connect(
    host='10.11.12.90\BDT',
    user='ibetlmarket',
    port=1433,
    password='m@rk3t2o15',
    database='MARKET'
)

import time
time.sleep(1)
gc.collect()

#                                                            REPORTE DE SEGUIMIENTO GESTION ALERTAS CARTERA

sql_ = """
	SELECT PERIODO, SEMAFORO, ESTADO, SEGMENTO_BE_AJUST, RESULTADO FROM ODS.HD_ALERTAS_CARTERA_SEGUIMIENTO_GESTION_CARTERA
"""
data = pd.read_sql(
   sql_ , connection_market
)
data['BANCA'] = 'BE'
data['RESULTADO'] = data['RESULTADO'].apply(lambda _: str(_).upper())
data['SEMAFORO'] = data['SEMAFORO'].apply(lambda _: str(_).upper())
data['ESTADO'] = data['ESTADO'].apply(lambda _: str(_).upper())
data.head(2)

### >> ALERTADOS (NO NEGROS)

subset = data[~data['SEMAFORO'].isin(['NEGRO'])]

pd.crosstab(subset['BANCA'], subset['PERIODO'])

### >> ALERTADOS POR ZONAL (NO NEGROS)

pd.crosstab(subset['SEGMENTO_BE_AJUST'], subset['PERIODO'])

### >> ALERTADOS POR GESTIONAR (POSTERIOR A LA DEPURACIÓN AUTOMÁTICA)


mismas_alertas = ['DEPURADO MISMAS ALERTAS EJECUTADO', 'DEPURADO MISMAS ALERTAS VALIDADO']
subset = data[
    (~data['SEMAFORO'].isin(['NEGRO', 'GRIS'])) & (~data['ESTADO'].isin(mismas_alertas))
]

pd.crosstab(subset['SEGMENTO_BE_AJUST'], subset['PERIODO'])

### >> REQUIERE GESTION

list_depura =['DEPURADO MISMAS ALERTAS EJECUTADO', 'DEPURADO MISMAS ALERTAS VALIDADO', 'EN PROCESO DE DEPURACION', 'DEPURADO']
subset = data[
    (~data['SEMAFORO'].isin(['NEGRO', 'GRIS'])) & (~data['ESTADO'].isin(list_depura))
]

pd.crosstab(subset['SEGMENTO_BE_AJUST'], subset['PERIODO'])

### >> GESTIONADOS

gestionado = ['EJECUTADO', 'VALIDADO', 'GUARDADO']
subset = data[
    (~data['SEMAFORO'].isin(['NEGRO', 'GRIS'])) & (data['ESTADO'].isin(gestionado))
]

pd.crosstab(subset['SEGMENTO_BE_AJUST'], subset['PERIODO'])

### >> RESULTADOS

subset = data[data['RESULTADO'].isin(['SEGUIMIENTO', 'PREOCUPA '])]

pd.crosstab(subset['SEGMENTO_BE_AJUST'], subset['PERIODO'])
